load-tests/requirements.txt
locust>=2.15.0
kafka-python>=2.0.2
